Move all bulk change stuff to whelktool module
