+++ librisworks/src/main/groovy/se/kb/libris/mergeworks/WorkComparator.groovy
@@ -1,6 +1,6 @@
 package se.kb.libris.mergeworks
 
-import whelk.util.DocumentComparator
+import datatool.util.DocumentComparator
 import se.kb.libris.mergeworks.compare.*
 
 import static Util.bestTitle
