+++ whelktool/src/test/groovy/datatool/bulkchange/FormBulkChangeSpec.groovy
@@ -1,14 +1,15 @@
+package datatool.bulkchange
+
 import spock.lang.Specification
-import whelk.BulkChange
 import whelk.Document
 
 import static whelk.util.Jackson.mapper
 
-class BulkChangeSpec extends Specification {
-    static List<Map> specs = BulkChangeSpec.class.getClassLoader()
-            .getResourceAsStream('bulk-change-specs.json')
-            .with { mapper.readValue(it, Map)['specs'] }
-    static BulkChange bulkChange2 = new BulkChange("", [:], ['contribution'] as Set)
+class FormBulkChangeSpec extends Specification {
+    static List<Map> specs = FormBulkChangeSpec.class.getClassLoader()
+            .getResourceAsStream('datatool/bulkchange/form-bulk-change-specs.json')
+            .with { mapper.readValue((InputStream) it, Map)['specs'] }
+    static FormBulkChange bulkChange = new FormBulkChange("", [:], [] as Set)
 
     def "get changed paths"() {
         given:
@@ -18,8 +19,8 @@ class BulkChangeSpec extends Specification {
         List expectedRemoved = spec['removedPaths']
 
         expect:
-        BulkChange.collectRemovedPaths(matchForm, targetForm) == expectedRemoved
-        BulkChange.collectAddedPaths(matchForm, targetForm) == expectedAdded
+        FormBulkChange.collectRemovedPaths(matchForm, targetForm) == expectedRemoved
+        FormBulkChange.collectAddedPaths(matchForm, targetForm) == expectedAdded
 
         where:
         spec << specs
@@ -29,12 +30,12 @@ class BulkChangeSpec extends Specification {
         given:
         Map before = (Map) Document.deepCopy(spec['matchForm'])
         Map after = (Map) Document.deepCopy(spec['targetForm'])
-        BulkChange.clearImplicitIds(before)
-        BulkChange.clearImplicitIds(after)
+        FormBulkChange.clearImplicitIds(before)
+        FormBulkChange.clearImplicitIds(after)
         Map formSpec = (Map) Document.deepCopy(spec.subMap(['matchForm', 'targetForm']))
 
         expect:
-        bulkChange2.modify(before, formSpec) == after
+        bulkChange.modify(before, formSpec) == after
 
         where:
         spec << specs
